fix: SupportResistance Decimal→f64 for Tauri IPC serialization
SupportResistance.supports/resistances/nearest_support/nearest_resistance
were Rust Decimal type, serializing as strings via Tauri IPC.
Changed to f64 for proper JSON number serialization.
Also fixes strategy.rs multiplication and data_fetcher mock data.

+++ crates/domain/src/lib.rs
@@ -279,10 +279,10 @@ impl Default for MovingAverage {
 #[derive(Debug, Clone, Serialize, Deserialize, Default, PartialEq)]
 pub struct SupportResistance {
     pub stock_id: String,
-    pub supports: Vec<Decimal>,
-    pub resistances: Vec<Decimal>,
-    pub nearest_support: Option<Decimal>,
-    pub nearest_resistance: Option<Decimal>,
+    pub supports: Vec<f64>,
+    pub resistances: Vec<f64>,
+    pub nearest_support: Option<f64>,
+    pub nearest_resistance: Option<f64>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
